backend_flask/requirements.txt
Flask
flask-pymongo
flask-cors
requests
python-dotenv

